requirements.txt
numpy==1.22.4
pandas==1.4.0
Pillow==9.3.0
streamlit==1.13.0
torch==1.13.0
torchvision==0.14.0
